Tweak to_float32

+++ ppocr/losses/rec_ctc_loss.py
@@ -29,7 +29,7 @@ class CTCLoss(nn.Layer):
     def forward(self, predicts, batch):
         if isinstance(predicts, (list, tuple)):
             predicts = predicts[-1]
-        predicts = predicts.transpose((1, 0, 2))
+        predicts = predicts.transpose((1, 0, 2)).astype("float32")
         N, B, _ = predicts.shape
         preds_lengths = paddle.to_tensor(
             [N] * B, dtype="int64", place=paddle.CPUPlace()
